fastcrypto_pq::sphincs

Module xmss

XMSS implementation from FIPS-205

Callers must ensure that the address input encodes XMSS tree address (layer & tree).

Structs

XmssParams

XmssSignature

Functions

xmss_node

Compute the XMSS subtree node at height z and index i (within that height).

xmss_pk_from_sig

xmss_sign

Type Aliases

XmssNode